+++ apps/src/lib/kits/maker/Button.js
@@ -1,7 +1,7 @@
 /** @file Wrapper around Johnny-Five Button component */
 import five from '@code-dot-org/johnny-five';
 import '../../../utils'; // For Function.prototype.inherits
-import {EXTERNAL_PINS} from './PlaygroundConstants';
+import {EXTERNAL_PINS} from './boards/circuitPlayground/PlaygroundConstants';
 
 /**
  * Wrap Johnny-Five's Button component to add attributes and customize behavior.
