+++ apps/src/lib/kits/maker/FakeBoard.js
@@ -1,7 +1,10 @@
 /** @file Fake for running Maker apps without an attached board. */
 import _ from 'lodash';
 import {EventEmitter} from 'events'; // provided by webpack's node-libs-browser
-import {J5_CONSTANTS, N_COLOR_LEDS} from './PlaygroundConstants';
+import {
+  J5_CONSTANTS,
+  N_COLOR_LEDS
+} from './boards/circuitPlayground/PlaygroundConstants';
 
 /**
  * Fake Maker Board for running Maker Toolkit apps without a board attached.
