+++ apps/src/lib/kits/maker/boards/circuitPlayground/PlaygroundConstants.js
@@ -1,4 +1,4 @@
-import experiments from '../../../util/experiments';
+import experiments from '../../../../../util/experiments';
 export const N_COLOR_LEDS = 10;
 export const EXTERNAL_PINS = [0, 1, 2, 3, 6, 9, 10, 12];
 export const TOUCH_PINS = [0, 2, 3, 6, 9, 10, 12];
