+++ apps/test/unit/lib/kits/maker/CircuitPlaygroundBoardTest.js
@@ -10,7 +10,7 @@ import CircuitPlaygroundBoard, {
 import {
   SONG_CHARGE,
   EXTERNAL_PINS
-} from '@cdo/apps/lib/kits/maker/PlaygroundConstants';
+} from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/PlaygroundConstants';
 import Led from '@cdo/apps/lib/kits/maker/Led';
 import {itImplementsTheMakerBoardInterface} from './MakerBoardTest';
 import experiments from '@cdo/apps/util/experiments';
