+++ apps/test/unit/lib/kits/maker/MakerBoardTest.js
@@ -2,7 +2,7 @@
 import sinon from 'sinon';
 import {EventEmitter} from 'events'; // see node-libs-browser
 import {expect} from '../../../../util/deprecatedChai';
-import {N_COLOR_LEDS} from '@cdo/apps/lib/kits/maker/PlaygroundConstants';
+import {N_COLOR_LEDS} from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/PlaygroundConstants';
 import MicroBitBoard from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitBoard';
 import CircuitPlaygroundBoard from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/CircuitPlaygroundBoard';
 import FakeBoard from '@cdo/apps/lib/kits/maker/FakeBoard';
