+++ apps/test/unit/lib/kits/maker/ThermometerTest.js
@@ -6,7 +6,7 @@ import Thermometer, {
 } from '@cdo/apps/lib/kits/maker/Thermometer';
 import Playground from 'playground-io';
 import {MicrobitStubBoard} from './makeStubBoard';
-import {sensor_channels} from '@cdo/apps/lib/kits/maker/MicroBitConstants';
+import {sensor_channels} from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitConstants';
 
 describe('Thermometer', function() {
   let testObj;
